Fix detection of the Start condition
Before 'check_start' would incorrectly handle situation where 'scl' and
'sda' fall in the same time slot. This lead to 'Start' detection when no
'Start' was present. This commit adds 'scl' check, so that if both 'sda'
and 'scl' go low and 'sda' falling edge trigger was first to fire,
'Start' is not reported. This change is similar in nature to the check
in the 'check_stop' method.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>